requirements.txt
numpy==1.19.0
spacy==2.2.4
pandas==1.0.1
Pillow==7.2.0
streamlit==0.63.0
tensorflow-cpu==2.2.0
tensorflow==2.2.0